Move my binary tree implementation into my "custom data structures" folder.
- This way, I can reuse the same binary tree data structure for other LeetCode problems that also use binary trees.
