+++ src/InterviewProblems/EasyQuestions/BFS/AverageOfLevelsInBinaryTree/AverageOfLevelsInBinaryTree.java
@@ -1,7 +1,11 @@
 package InterviewProblems.EasyQuestions.BFS.AverageOfLevelsInBinaryTree;
 
+import DataStructures.CustomDataStructures.BinaryTree.BinaryTreeNode;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Queue;
 
 public class AverageOfLevelsInBinaryTree {
